Print unhandeled opcodes as uint8_t (#633)
* Update gfx_pc.cpp

* Update gfx_pc.cpp

+++ src/graphic/Fast3D/gfx_pc.cpp
@@ -3828,7 +3828,8 @@ const char* GfxGetOpcodeName(int8_t opcode) {
         if (ucode_handlers[ucode_handler_index]->contains(opcode)) {
             return ucode_handlers[ucode_handler_index]->at(opcode).first;
         } else {
-            SPDLOG_CRITICAL("Unhandled OP code: 0x{:X}, for loaded ucode: {}", opcode, (uint32_t)ucode_handler_index);
+            SPDLOG_CRITICAL("Unhandled OP code: 0x{:X}, for loaded ucode: {}", (uint8_t)opcode,
+                            (uint32_t)ucode_handler_index);
             return nullptr;
         }
     }
@@ -3868,7 +3869,8 @@ static void gfx_step() {
                 return;
             }
         } else {
-            SPDLOG_CRITICAL("Unhandled OP code: 0x{:X}, for loaded ucode: {}", opcode, (uint32_t)ucode_handler_index);
+            SPDLOG_CRITICAL("Unhandled OP code: 0x{:X}, for loaded ucode: {}", (uint8_t)opcode,
+                            (uint32_t)ucode_handler_index);
         }
     }
